feat: show notification on robot delete with existing runs

+++ src/api/storage.ts
@@ -5,6 +5,11 @@ import { ScheduleSettings } from "../components/molecules/ScheduleSettings";
 import { CreateRunResponse, ScheduleRunResponse } from "../pages/MainPage";
 import { apiUrl } from "../apiConfig";
 
+
+
+
+
+
 export const getStoredRecordings = async (): Promise<string[] | null> => {
   try {
     const response = await axios.get(`${apiUrl}/storage/recordings`);
@@ -77,18 +82,49 @@ export const getStoredRecording = async (id: string) => {
   }
 }
 
+
+
+export const checkRunsForRecording = async (id: string): Promise<boolean> => {
+    
+  
+  try {
+    const response = await axios.get(`${apiUrl}/storage/recordings/${id}/runs`);
+
+    const runs = response.data;
+    console.log(runs.runs.totalCount)
+    return runs.runs.totalCount > 0;
+  } catch (error) {
+    console.error('Error checking runs for recording:', error);
+    return false;
+  }
+};
+
+
 export const deleteRecordingFromStorage = async (id: string): Promise<boolean> => {
+  
+  const hasRuns = await checkRunsForRecording(id);
+  
+  if (hasRuns) {
+    
+    return false;
+  }
   try {
     const response = await axios.delete(`${apiUrl}/storage/recordings/${id}`);
     if (response.status === 200) {
-      return response.data;
+      
+      return true;
     } else {
       throw new Error(`Couldn't delete stored recording ${id}`);
     }
   } catch (error: any) {
     console.log(error);
+    
     return false;
   }
+
+  
+
+  
 };
 
 export const deleteRunFromStorage = async (id: string): Promise<boolean> => {
@@ -123,7 +159,7 @@ export const createRunForStoredRecording = async (id: string, settings: RunSetti
   try {
     const response = await axios.put(
       `${apiUrl}/storage/runs/${id}`,
-      { ...settings });
+      { ...settings }); 
     if (response.status === 200) {
       return response.data;
     } else {
